Change unit and type for portion sizes

+++ src/main/java/Challenge1.java
@@ -3,10 +3,10 @@ public class Challenge1 {
      * Given an array of arrays of portion sizes, return the index with the greatest average portion size.
      *
      * @param restaurants A 2D array such that each element of the outer array is
-     *                      an array representing the portion sizes of a restaurant's entrees (in ounces).
+     *                      an array representing the portion sizes of a restaurant's entrees (in grams).
      * @return the index of the array with greatest average portion size.
      */
-    public int largestPortions(double[][] restaurants) {
+    public int largestPortions(int[][] restaurants) {
         return 0;
     }
 }
